update binwin64 (probably no different from previous one) add Stacking Fault-III to tutorial list fixup HDF5 importer for images - now recovers wavelength, distance, pixel size, etc. fix float/integer incompatibility in Flat Bkg values enable/disable if calibrant/nocalibrant for both Calibrate & Recalibrate fix a float/int error in mouse positioning on image
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@2284 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ GSASIIplot.py
@@ -3945,8 +3945,8 @@ def PlotImage(G2frame,newPlot=False,event=None,newImage=True):
             else:
                 xpos = event.xdata
                 ypos = event.ydata
-                xpix = xpos*scalex
-                ypix = ypos*scaley
+                xpix = int(xpos*scalex)
+                ypix = int(ypos*scaley)
                 Int = 0
                 if (0 <= xpix <= sizexy[0]) and (0 <= ypix <= sizexy[1]):
                     Int = G2frame.ImageZ[ypix][xpix]
